Convert Cards to Functional Components.
This was a little harder. Functional components don't correctly inherit
their parent's scope IDs for scoped styles, so I had to undo some of
that. Also had to adjust the way click events work so it can use the
basic click event, instead of a special one defined on the card.